refactor: PathVertexStation 패키지 이동
