+++ src/main/java/nextstep/subway/path/application/PathFinder.java
@@ -3,7 +3,7 @@ package nextstep.subway.path.application;
 import nextstep.subway.line.domain.Line;
 import nextstep.subway.line.domain.LineRepository;
 import nextstep.subway.line.domain.Section;
-import nextstep.subway.path.domain.PathVertexStation;
+import nextstep.subway.path.dto.PathVertexStation;
 import nextstep.subway.path.dto.PathResponse;
 import nextstep.subway.station.domain.Station;
 import nextstep.subway.station.domain.StationRepository;
